refactor: step별 컴포넌트 분리
